fixed seeder

+++ app/seeds/users.py
@@ -1,4 +1,5 @@
 from app.models import db, User
+import os
 
 
 # Adds a demo user, you can add other users here if you want
@@ -23,7 +24,8 @@ def seed_users():
 # sqlite3 in development you need to instead use DELETE to remove all data and
 # it will reset the primary keys for you as well.
 def undo_users():
-    db.session.execute(f"TRUNCATE table users RESTART IDENTITY CASCADE;")
-    db.session.execute("DELETE FROM users")
-
+    if os.environ.get("FLASK_ENV") == 'development':
+        db.session.execute('DELETE FROM users;')
+    else:
+        db.session.execute('TRUNCATE users RESTART IDENTITY CASCADE;')
     db.session.commit()
